+++ iris/service/content/file/__init__.py
@@ -2,6 +2,6 @@ from .document import File  # noqa
 
 
 def includeme(config):
-    config.add_route('blob_api',
-                     '/blobs',
+    config.add_route('file_api',
+                     '/files',
                      static=True)