+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changes For IRIS Service
 unreleased
 ==========
 
- - added blob service (upload and download)
+ - added file service (upload and download)
    DEPLOY:
     - recreate files table in Crate
     - enable libmagic in Docker containers